+++ server/routes/api/mail.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const mailController = require("../../controller/api/mail");
+
+router.get("/submit", mailController.toProfessor);
+
+module.exports = router;
